requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
h11==0.14.0
idna==3.10
influxdb-client==1.48.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
